Tests for utils.
* migrate utils tests to new `utils_test.dart`
* extract test utils
* add tests for untested utils functions
