+++ test/rule_test.dart
@@ -19,7 +19,6 @@ import 'package:linter/src/analyzer.dart';
 import 'package:linter/src/ast.dart';
 import 'package:linter/src/formatter.dart';
 import 'package:linter/src/rules.dart';
-import 'package:linter/src/utils.dart';
 import 'package:linter/src/rules/implementation_imports.dart';
 import 'package:linter/src/rules/package_prefixed_library_names.dart';
 import 'package:path/path.dart' as p;
@@ -147,135 +146,6 @@ defineRuleUnitTests() {
       testEach(good, isValidDartIdentifier, isTrue);
       var bad = ['if', '42', '3', '2f'];
       testEach(bad, isValidDartIdentifier, isFalse);
-    });
-    group('pubspec', () {
-      testEach(['pubspec.yaml', '_pubspec.yaml'], isPubspecFileName, isTrue);
-      testEach(['__pubspec.yaml', 'foo.yaml'], isPubspecFileName, isFalse);
-    });
-
-    group('camel case', () {
-      group('upper', () {
-        var good = [
-          '_FooBar',
-          'FooBar',
-          '_Foo',
-          'Foo',
-          'F',
-          'FB',
-          'F1',
-          'FooBar1',
-          '\$Foo',
-          'Bar\$',
-          'Foo\$Generated',
-          'Foo\$Generated\$Bar'
-        ];
-        testEach(good, isCamelCase, isTrue);
-        var bad = ['fooBar', 'foo', 'f', '_f', 'F_B'];
-        testEach(bad, isCamelCase, isFalse);
-      });
-    });
-    group('library prefixes', () {
-      var good = [
-        'foo_bar',
-        'foo',
-        'foo_bar_baz',
-        'p',
-        'p1',
-        'p21',
-        'p1ll0',
-        '_foo',
-        r'$foo',
-        '__foo',
-        r'$_foo',
-        r'$__foo',
-      ];
-      testEach(good, isValidLibraryPrefix, isTrue);
-
-      var bad = [
-        'JSON',
-        'JS',
-        'Math',
-        'jsUtils',
-        'foo_Bar',
-        'F_B',
-        '1',
-        '1b',
-        r'foo$',
-        r'f$oo',
-        r'$$foo',
-        r'$_$foo',
-        r'_$foo',
-      ];
-      testEach(bad, isValidLibraryPrefix, isFalse);
-    });
-
-    group('lower_case_underscores', () {
-      var good = ['foo_bar', 'foo', 'foo_bar_baz', 'p', 'p1', 'p21', 'p1ll0'];
-      testEach(good, isLowerCaseUnderScore, isTrue);
-
-      var bad = [
-        'Foo',
-        'fooBar',
-        'foo_Bar',
-        'foo_',
-        '_f',
-        'F_B',
-        'JS',
-        'JSON',
-        '1',
-        '1b',
-      ];
-      testEach(bad, isLowerCaseUnderScore, isFalse);
-    });
-    group('qualified lower_case_underscores', () {
-      var good = [
-        'bwu_server.shared.datastore.some_file',
-        'foo_bar.baz',
-        'foo_bar',
-        'foo.bar',
-        'foo_bar_baz',
-        'foo',
-        'foo_',
-        'foo.bar_baz.bang',
-        //See: https://github.com/flutter/flutter/pull/1996
-        'pointycastle.impl.ec_domain_parameters.gostr3410_2001_cryptopro_a',
-        'a.b',
-        'a.b.c',
-        'p2.src.acme'
-      ];
-      testEach(good, isLowerCaseUnderScoreWithDots, isTrue);
-
-      var bad = ['Foo', 'fooBar.', '.foo_Bar', '_f', 'F_B', 'JS', 'JSON'];
-      testEach(bad, isLowerCaseUnderScoreWithDots, isFalse);
-    });
-    group('lowerCamelCase', () {
-      var good = [
-        'fooBar',
-        'foo',
-        'f',
-        'f1',
-        '_f',
-        '_foo',
-        '_',
-        'F',
-        '__x',
-        '___x',
-        '\$foo',
-        'bar\$',
-        'foo\$Generated',
-        'foo\$Generated\$Bar'
-      ];
-      testEach(good, isLowerCamelCase, isTrue);
-
-      var bad = ['Foo', 'foo_', 'foo_bar', '_X'];
-      testEach(bad, isLowerCamelCase, isFalse);
-    });
-    group('isUpperCase', () {
-      var caps = new List<int>.generate(26, (i) => 'A'.codeUnitAt(0) + i);
-      testEachInt(caps, isUpperCase, isTrue);
-
-      var bad = ['a', '1', 'z'].map((c) => c.codeUnitAt(0));
-      testEachInt(bad, isUpperCase, isFalse);
     });
     group('libary_name_prefixes', () {
       testEach(
